Correcao desafio

+++ src/main/java/com/devsuperior/dsmeta/dto/SumarioVendasDTO.java
@@ -4,13 +4,13 @@ import com.devsuperior.dsmeta.projections.SumarioVendasProjection;
 
 public class SumarioVendasDTO {
 	private String name;
-	private Long sum;
+	private Double sum;
 	
 	
 	public SumarioVendasDTO() {
 		
 	}
-	public SumarioVendasDTO(String name, Long sum) {
+	public SumarioVendasDTO(String name, Double sum) {
 		this.name = name;
 		this.sum = sum;
 	}
@@ -23,13 +23,13 @@ public class SumarioVendasDTO {
 	public String getName() {
 		return name;
 	}
-	public Long getSum() {
+	public Double getSum() {
 		return sum;
 	}
 	public void setName(String name) {
 		this.name = name;
 	}
-	public void setSum(Long sum) {
+	public void setSum(Double sum) {
 		this.sum = sum;
 	}
