+++ src/main/java/com/devsuperior/dsmeta/projections/SumarioVendasProjection.java
@@ -2,6 +2,6 @@ package com.devsuperior.dsmeta.projections;
 
 public interface SumarioVendasProjection {
 	public String getName();
-	public Long getSum(); 
+	public Double getSum(); 
 
 }
